Finalizada revisão das migrations do geral e do academico

+++ modulos/Geral/Database/Migrations/2016_07_28_170700_create_gra_titulacoes_table.php
@@ -16,7 +16,7 @@ class CreateGraTitulacoesTable extends Migration
             $table->increments('tit_id');
             $table->string('tit_nome', 60);
             $table->string('tit_descricao', 150)->nullable();
-            $table->integer('tit_peso');
+            $table->smallInteger('tit_peso');
 
             $table->timestamps();
         });
